+++ lib/push_x/application.ex
@@ -21,21 +21,27 @@ defmodule PushX.Application do
              size: PushX.Config.finch_pool_size(),
              count: PushX.Config.finch_pool_count(),
              protocols: [:http2],
-             conn_opts: [transport_opts: [timeout: PushX.Config.connect_timeout()]]
+             conn_opts: [
+               transport_opts: [timeout: PushX.Config.connect_timeout(), keepalive: true]
+             ]
            ],
            # APNS Sandbox
            "https://api.sandbox.push.apple.com" => [
              size: PushX.Config.finch_pool_size(),
              count: PushX.Config.finch_pool_count(),
              protocols: [:http2],
-             conn_opts: [transport_opts: [timeout: PushX.Config.connect_timeout()]]
+             conn_opts: [
+               transport_opts: [timeout: PushX.Config.connect_timeout(), keepalive: true]
+             ]
            ],
            # FCM (Firebase Cloud Messaging)
            "https://fcm.googleapis.com" => [
              size: PushX.Config.finch_pool_size(),
              count: PushX.Config.finch_pool_count(),
              protocols: [:http2],
-             conn_opts: [transport_opts: [timeout: PushX.Config.connect_timeout()]]
+             conn_opts: [
+               transport_opts: [timeout: PushX.Config.connect_timeout(), keepalive: true]
+             ]
            ],
            :default => [
              size: PushX.Config.finch_pool_size(),